Fix merging error

+++ src/Mod/PartDesign/App/FeatureSketchBased.cpp
@@ -1345,13 +1345,9 @@ void ProfileBased::handleChangedPropertyName(
 
 bool ProfileBased::isElementGenerated(const TopoShape &shape, const Data::MappedName &name) const
 {
-    // auto profile = Profile.getValue();
-    // if (!profile)
-    //     return FeatureAddSub::isElementGenerated(shape, name);
-    // long profileTag = profile->getID();
     bool res = false;
     shape.traceElement(name,
-        [&] (const Data::MappedName &, size_t, long tag2, long) {
+        [&] (const Data::MappedName &, int, long tag2, long) {
             if (std::abs(tag2) == this->getID()) {
                 res = true;
                 return true;
